opening the graphql playground on default browser in dev mode implemented.

+++ src/resolvers/block.ts
@@ -2,6 +2,7 @@ import { Arg, Ctx, Mutation, Resolver, UseMiddleware } from "type-graphql";
 import { getConnection } from "typeorm";
 import { Block } from "../entities/Block";
 import { Follow } from "../entities/Follow";
+import { User } from "../entities/User";
 import { isAuth } from "../middleware/isAuth";
 import { MyContext } from "../types";
 
@@ -17,6 +18,9 @@ export class BlockResolver {
       where: { userId, blockedByUserId: myUserId },
     });
     if (blocked) return true;
+    const user = await User.findOne(userId);
+    if (!user) return false;
+    if (user.amIDeactivated) return false;
     await getConnection().transaction(async (em) => {
       em.insert(Block, { userId, blockedByUserId: myUserId });
       em.delete(Follow, { userId, followerId: myUserId });
@@ -35,6 +39,9 @@ export class BlockResolver {
       where: { userId, blockedByUserId: myUserId },
     });
     if (!blocked) return true;
+    const user = await User.findOne(userId);
+    if (!user) return false;
+    if (user.amIDeactivated) return false;
     await Block.delete({ userId, blockedByUserId: myUserId });
     return true;
   }
